修正 codemirror 字型與 課程大綱 儲存 bug

+++ grails-app/controllers/codecanaan/CourseController.groovy
@@ -218,7 +218,9 @@ class CourseController {
             else {
                 log.info "Remove ${course.title} from Open Course List"
                 def oc = OpenCourse.findByCourse(course)
-                oc.delete(flush: true)
+                if (oc) {
+                    oc.delete(flush: true)
+                }
             }
         }
